updated Book component to add books to savedBooks
