+++ client/src/index.css
@@ -1,5 +1,8 @@
+@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@100;400;700&display=swap");
+
 body {
-  font-family: 'Rubik', sans-serif;
+  font-family: "Roboto", sans-serif;
+  background-color: lightgray;
 }
 
 h1,
@@ -8,7 +11,7 @@ h3,
 h4,
 h5,
 h6 {
-  font-family: 'Rubik Mono One', sans-serif;
+  font-family: "Roboto", sans-serif;
 }
 
 .small {
