+++ client/src/pages/SearchBooks.js
@@ -65,29 +65,7 @@ const SearchBooks = () => {
     }
   };
 
-  // create function to handle saving a book to our database
-  const handleSaveBook = async (bookId) => {
-    // find the book in `searchedBooks` state by the matching id
-    const bookToSave = searchedBooks.find((book) => book.bookId === bookId);
 
-    // get token
-    const token = Auth.loggedIn() ? Auth.getToken() : null;
-
-    if (!token) {
-      return false;
-    }
-
-    try {
-      await saveBook({
-        variables: { input: bookToSave },
-      });
-
-      // if book successfully saves to user's account, save book id to state
-      setSavedBookIds([...savedBookIds, bookToSave.bookId]);
-    } catch (err) {
-      console.error(err);
-    }
-  };
 
   return (
     <>
@@ -128,6 +106,12 @@ const SearchBooks = () => {
             <Book
               key={book.bookId}
               book={book}
+              searchedBooks={searchedBooks}
+              setSearchedBooks={setSearchedBooks} 
+              saveBook={saveBook}
+              savedBookIds={savedBookIds}
+              setSavedBookIds={setSavedBookIds}
+
             />
           );
         })}
