+++ package.json
@@ -21,6 +21,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^16.0.3",
     "if-env": "^1.0.4"
   },
   "devDependencies": {
