+++ server/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "apollo-server-express": "^2.19.1",
     "bcrypt": "^4.0.1",
+    "dotenv": "^16.0.3",
     "express": "^4.17.1",
     "faker": "^4.1.0",
     "jsonwebtoken": "^8.5.1",
