+++ server/server.js
@@ -3,6 +3,7 @@ const express = require("express");
 const { ApolloServer } = require("apollo-server-express");
 const { authMiddleware } = require("./utils/auth");
 
+require('dotenv').config()
 const path = require("path");
 const db = require("./config/connection");
